project done using numpy library
Ratings of the survey has been categorised into 3 labels: Detractors: Respondents with a score of 0-6 , Passive: Respondents with a score of 7-8, Promoters: Respondents with a score of 9-10. NPS helps a brand in gauging its brand value and sentiment in the market. Promoters are highly likely to recommend the product or sevice. Hence, bringing in more business. whereas, Detractors are likely to recommend against the product or service’s usage. Hence, bringing the business down. 

Here, we are making a biased calculation to compute NPS score as in Detractors' scoring range is higher than the Promoters' scoring range. Because our target is observe the number of people promoting  as well as detracting the business. The greater the promotes' count and the lesser the detractors count, the higher the NPS percentage will be, the better the busines will grow.

As we see here in this case study the NPS percentage is found 23.7% which is way too lower than 70% ( we assume an optimum NPS score for growing a business) where promoters' percentage is 28% and detractors' percentage is 52%. So we can conclude that the business is not at all doing good and the BoD probably should dissolve the business.